[presto-orc] Remove code duplication in StripeReader
